First project in the nanodegree about US bikeshare data
This is the first project in the nanodegree. It deals with a reduced dataset of bikeshare data of three US cities. It calculates varous metrics such as the most busy stations, days and months of usage, as well as average trip time. In addition to some statistics about the users.